requirements.txt
black==23.3.0
blinker==1.6.2
click==8.1.3
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.7.0
psycopg2==2.9.6
pyscopg2==66.0.2
python-dotenv==1.0.0
Werkzeug==2.3.6
